I've refactored the Supabase Edge Functions to be self-contained, which should resolve the critical `Module not found` deployment error you were seeing. My previous attempt using a shared module for logging was incompatible with the Supabase deployment environment, so I took a new approach.
Here are the key changes I made to the backend:
*   I removed the `_shared` directory.
*   I copied the logging helper functions (`logToolRun`, `updateToolRun`) directly into each of the seven affected Edge Functions (like `schema-generator`, `citation-tracker`, `ai-visibility-audit`, etc.).
*   I made the logic within those helper functions more robust and corrected them to use the proper database column names (`output_payload`, `error_message`).
*   I updated the function calls to match the new helper signatures.

I also included the frontend fixes from my previous work:
*   I ensured the `projectId` is now correctly passed from the frontend to the `ai-content-generator` and `prompt-match-suggestions` services.
*   I added a missing `apiService` import in `OptimizationPlaybooks.tsx`.
*   I fixed a CORS preflight request issue in `get-integrations`.

Please note: The local development server environment appears to be broken, as it was failing to load the application even before my changes. Because of this, I couldn't verify this fix locally, but the changes directly address the deployment error you reported.
