+++ supabase/functions/ai-visibility-audit/index.ts
@@ -5,9 +5,50 @@ const corsHeaders = {
   'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
 };
 
-import { logToolRun, updateToolRun } from '../_shared/dbLogger.ts';
-
 // --- Type Definitions & Helpers ---
+
+async function logToolRun(supabase, projectId, toolName, inputPayload) {
+  if (!projectId) {
+    throw new Error("logToolRun error: projectId is required.");
+  }
+  const { data, error } = await supabase
+    .from("tool_runs")
+    .insert({
+      project_id: projectId,
+      tool_name: toolName,
+      input_payload: inputPayload,
+      status: "running",
+    })
+    .select("id")
+    .single();
+
+  if (error) {
+    console.error("Error logging tool run:", error);
+    throw new Error(`Failed to log tool run. Supabase error: ${error.message}`);
+  }
+  if (!data || !data.id) {
+    console.error("No data or data.id returned from tool_runs insert.");
+    throw new Error("Failed to log tool run: No data returned after insert.");
+  }
+  return data.id;
+}
+
+async function updateToolRun(supabase, runId, status, outputPayload, errorMessage) {
+  if (!runId) {
+    console.error("updateToolRun error: runId is required.");
+    return;
+  }
+  const update = {
+    status,
+    completed_at: new Date().toISOString(),
+    output_payload: errorMessage ? { error: errorMessage } : outputPayload || null,
+    error_message: errorMessage || null,
+  };
+  const { error } = await supabase.from("tool_runs").update(update).eq("id", runId);
+  if (error) {
+    console.error(`Error updating tool run ID ${runId}:`, error);
+  }
+}
 
 async function updateToolRunProgress(supabase: SupabaseClient, runId: string, progress: Record<string, unknown>) {
   const { error } = await supabase.from('tool_runs').update({ progress }).eq('id', runId);
@@ -112,7 +153,7 @@ export const auditService = async (req: Request, supabase: SupabaseClient): Prom
     const scores = Object.values(finalResult.subscores);
     finalResult.overallScore = Math.round(scores.reduce((a, b) => a + b, 0) / scores.length);
 
-    await updateToolRun(supabase, runId, 'completed', finalResult);
+    await updateToolRun(supabase, runId, 'completed', finalResult, null);
 
     return new Response(JSON.stringify({ success: true, data: { runId, ...finalResult } }), {
       headers: { ...corsHeaders, "Content-Type": "application/json" },
